old_backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fiona>=1.10.1",
    "geopandas>=1.1.1",
    "numpy>=2.3.5",
    "opencv-python-headless>=4.11.0.86",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "shapely>=2.1.2",
    "tqdm>=4.67.1",
    "torch>=2.2.0",
    "transformers>=4.38.0",
    "pytorch-lightning>=2.1.0",
    "evaluate>=0.4.1",
    "scikit-learn>=1.4.0",
    "albumentations>=2.0.8",
    "typer>=0.20.0",
    "pillow>=12.0.0",
    "buildingregulariser>=0.2.4",
    "pyproj>=3.7.2",
    "fastapi>=0.123.4",
    "uvicorn>=0.38.0",
    "pvlib>=0.13.1",
    "pandas>=2.3.3",
    "gdal==3.10.3",
    "rvt-py>=2.2.3",
    "openmeteo-requests>=1.7.4",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "segment-geospatial>=0.14.1",
    "leafmap>=0.57.9",
    "localtileserver>=0.10.6",
]
